+++ lib/rspec/mocks/example_methods.rb
@@ -128,7 +128,15 @@ module RSpec
       #   # You can also use most message expectations:
       #   expect(invitation).to have_received(:accept).with(mailer).once
       def have_received(method_name)
-        HaveReceived.new(method_name)
+        Matchers::HaveReceived.new(method_name)
+      end
+
+      def self.included(klass)
+        klass.class_eval do
+          # This gets mixed in so that if `RSpec::Matchers` is included in
+          # `klass` later, it's definition of `expect` will take precedence.
+          include ExpectHost unless method_defined?(:expect)
+        end
       end
 
     private
@@ -139,6 +147,10 @@ module RSpec
         RSpec::Mocks::Mock.new(*args)
       end
 
+      # This module exists to host the `expect` method for cases where
+      # rspec-mocks is used w/o rspec-expectations.
+      module ExpectHost
+      end
     end
   end
 end