+++ spec/rspec/mocks/null_object_mock_spec.rb
@@ -103,5 +103,14 @@ module RSpec
         expect(obj).not_to be_null_object
       end
     end
+
+    describe "when using the :wrapped syntax" do
+      include_context "with syntax", :wrapped
+
+      it 'still supports null object doubles' do
+        obj = double("foo").as_null_object
+        expect(obj.foo.bar.bazz).to be(obj)
+      end
+    end
   end
 end